Merged in webrtc-rtt-should-always-be-measured (pull request #160)
WebRTC RTT should always be measured.

In the past we needed to turn on and off the measurement, and to start it after other measurements, but this is no more the case.

There are no cases where we do NOT want to measure it, and we don’t need the delay any more.

Approved-by: Daniel Fredriksson