Merged in SG-2083-find-the-absolute-delay-using-fr (pull request #192)
Add stats for measuring frame delay

The frame delay is the time between a frame being captured from the
emulator and the client showing it. This can be used as a complement
to touchRTT, to get more consistent and frequent measurements of
a big part of the experienced “input lag”.

This frame delay cannot be calculated directly in the SDK, since not all
information is known at the beginning of the stream. The stats added
in this PR can instead be used to calculate it in BigQuery later.

We can do this by using the recently added stats:

* latestFrameRtpTimestamp - The timestamp encoded in the latest frame.
   This is given relative to a random point in time

* latestFrameClientTimestamp - The time when the client received/decoded
   that frame

in combination with the stats added in this PR:

* estimatedPlayoutTimestamp - The estimated time that the currently played
   frame was captured, according to the server’s clock. This is only available
   after a few seconds.

* estimatedTimeOffset - The estimated offset between the client’s and
   server’s system clocks. We measure this ourselves, and multiple
   measurements are needed to get enough accuracy.

To get the frame delay from these, we can:

1. Translate latestFrameRtpTimestamp into timestamps according to the server
   clock, by comparing it to estimatedPlayoutTimestamp at some point where
   the latter is available.

2. Translate this into timestamps according to the client clock, by using
   estimatedTimeOffset.

3. Subtract this from latestFrameClientTimestamp to get the frame delay.

This also relies on these PRs:
https://bitbucket.org/appland/streaming-games/pull-requests/871
https://bitbucket.org/appland/streaming-games/pull-requests/872

Approved-by: Francesco Orsenigo