Record individual timestamps of each frame (pull request #219)
Save all timestamps of a frame being ready to display, and send
them to the backend with the other measurements.

This uses APIs for modifying video (currently only implemented
in Chrome), to insert a callback after a frame is decoded but before
it's rendered. Instead of modifying the frame, we only record the
current time.

Since this information is available in EmulatorWebrtcView rather than JsepProtocol, move the handling of StreamingEvent.REQUEST_WEB_RTC_MEASUREMENT up a level.

Approved-by: yurii.shypulin
Approved-by: Francesco Orsenigo